requirements.txt
langchain>=0.3.0,<0.4.0
langgraph>=0.2.0,<0.3.0
langgraph-api>=0.0.12,<0.1.0
langchain-core>=0.3.23
langchain-community>=0.3.20
#langchain-openai
langchain-groq>=0.2.4
trustcall
ipython
scipy
pygame
groq 
sounddevice==0.5.2
streamlit
streamlit_chat_widget
elevenlabs
ffmpeg
python-dotenv
streamlit_float



